requirements.txt
fastapi>=0.110.0,<1.0.0
uvicorn[standard]>=0.27.0,<1.0.0
sqlalchemy>=2.0.0,<3.0.0
psycopg2-binary>=2.9.9,<3.0.0
alembic>=1.13.0,<2.0.0
pydantic-settings>=2.2.0,<3.0.0
passlib[bcrypt]>=1.7.4,<2.0.0
python-jose[cryptography]>=3.3.0,<4.0.0
